+++ jasentool/cli.py
@@ -26,36 +26,36 @@ def __query(group, required):
     group.add_argument('-q', '--query', required=required, nargs='+', help='sample query')
 
 def __sample_id(group, required):
-    """Add sample_id argument to group"""
-    group.add_argument('--sample_id', required=required, type=str, help='sample ID')
+    """Add sample-id argument to group"""
+    group.add_argument('--sample-id', required=required, type=str, help='sample ID')
 
 def __input_dir(group, required, help):
-    """Add input_dir argument to group"""
-    group.add_argument('--input_dir', required=required, help=help)
+    """Add input-dir argument to group"""
+    group.add_argument('--input-dir', required=required, help=help)
 
 def __input_file(group, required, help):
-    """Add input_file argument to group"""
-    group.add_argument('-i', '--input_file', required=required, nargs='+', help=help)
+    """Add input-file argument to group"""
+    group.add_argument('-i', '--input-file', required=required, nargs='+', help=help)
 
 def __csv_file(group, required, help):
-    """Add csv_file argument to group"""
-    group.add_argument('--csv_file', required=required, help=help)
+    """Add csv-file argument to group"""
+    group.add_argument('--csv-file', required=required, help=help)
 
 def __sh_file(group, required, help):
-    """Add sh_file argument to group"""
-    group.add_argument('--sh_file', required=required, default=None, help=help)
+    """Add sh-file argument to group"""
+    group.add_argument('--sh-file', required=required, default=None, help=help)
 
 def __bam_file(group, required):
-    """Add bam_file argument to group"""
-    group.add_argument('--bam_file', required=required, type=str, help='input bam file')
+    """Add bam-file argument to group"""
+    group.add_argument('--bam-file', required=required, type=str, help='input bam file')
 
 def __bed_file(group, required):
-    """Add bed_file argument to group"""
-    group.add_argument('--bed_file', required=required, type=str, help='input bed file')
+    """Add bed-file argument to group"""
+    group.add_argument('--bed-file', required=required, type=str, help='input bed file')
 
 def __baits_file(group, required):
-    """Add baits_file argument to group"""
-    group.add_argument('--baits_file', required=required, type=str, default=None,
+    """Add baits-file argument to group"""
+    group.add_argument('--baits-file', required=required, type=str, default=None,
                        help='input baits file')
 
 def __reference(group, required, help):
@@ -63,39 +63,39 @@ def __reference(group, required, help):
     group.add_argument('--reference', required=required, type=str, help=help)
 
 def __output_file(group, required, help):
-    """Add output_file argument to group"""
-    group.add_argument('-o', '--output_file', required=required, type=str, help=help)
+    """Add output-file argument to group"""
+    group.add_argument('-o', '--output-file', required=required, type=str, help=help)
 
 def __output_dir(group, required):
-    """Add output_dir argument to group"""
-    group.add_argument('--output_dir', required=required, type=str,
+    """Add output-dir argument to group"""
+    group.add_argument('--output-dir', required=required, type=str,
                        help='directory to output files')
 
 def __analysis_dir(group, required):
-    """Add analysis_dir argument to group"""
-    group.add_argument('--analysis_dir', required=required, type=str,
+    """Add analysis-dir argument to group"""
+    group.add_argument('--analysis-dir', required=required, type=str,
                        help='analysis results dir containing jasen results')
 
 def __restore_dir(group, required):
-    """Add restore_dir argument to group"""
-    group.add_argument('--restore_dir', required=required, type=str,
+    """Add restore-dir argument to group"""
+    group.add_argument('--restore-dir', required=required, type=str,
                        default='/fs2/seqdata/restored',
                        help='directory user wishes spring files to be restored to')
 
 def __remote_dir(group, required):
-    """Add remote_dir argument to group"""
-    group.add_argument('--remote_dir', required=required, type=str,
+    """Add remote-dir argument to group"""
+    group.add_argument('--remote-dir', required=required, type=str,
                        default='/fs1/bjorn/jasen',
                        help='directory user wishes spring files to be restored to')
 
 def __restore_file(group, required):
-    """Add restore_file argument to group"""
-    group.add_argument('--restore_file', required=required, type=str,
+    """Add restore-file argument to group"""
+    group.add_argument('--restore-file', required=required, type=str,
                        help='filepath bash shell script (.sh) to be output')
 
 def __missing_log(group, required):
-    """Add missing_log argument to group"""
-    group.add_argument('--missing_log', required=required, type=str,
+    """Add missing-log argument to group"""
+    group.add_argument('--missing-log', required=required, type=str,
                        default='missing_samples.log',
                        help='file containing missing files')
 
@@ -119,20 +119,20 @@ def __uri(group):
                         Use: `sudo lsof -iTCP -sTCP:LISTEN | grep mongo` to get address')
 
 def __db_name(group, required):
-    """Add db_name argument to group"""
-    group.add_argument('--db_name', required=required,
+    """Add db-name argument to group"""
+    group.add_argument('--db-name', required=required,
                        help='Mongodb database name address. \
                         Use: `show dbs` to get db name')
 
 def __db_collection(group, required):
-    """Add db_collection argument to group"""
-    group.add_argument('--db_collection', required=required,
+    """Add db-collection argument to group"""
+    group.add_argument('--db-collection', required=required,
                        help='Mongodb collection name. \
                         Use: `show collections` to get db collection')
 
 def __out_format(group, required):
-    """Add out_format argument to group"""
-    group.add_argument('-f', '--out_format', required=required, type=str,
+    """Add out-format argument to group"""
+    group.add_argument('-f', '--out-format', required=required, type=str,
                        default="bed", help='output format')
 
 def __accession(group, required):
@@ -140,8 +140,8 @@ def __accession(group, required):
     group.add_argument('-a', '--accession', required=required, type=str, help='accession number')
 
 def __remote_hostname(group, required):
-    """Add remote_hostname argument to group"""
-    group.add_argument('--remote_hostname', required=required, type=str,
+    """Add remote-hostname argument to group"""
+    group.add_argument('--remote-hostname', required=required, type=str,
                        default='rs-fe1.lunarc.lu.se', help='remote hostname')
 
 def __prefix(group):
@@ -150,8 +150,8 @@ def __prefix(group):
                        help='prefix for all output files')
 
 def __auto_start(group, required):
-    """Add auto_start argument to group"""
-    group.add_argument('--auto_start', required=required, dest='auto_start', action='store_true',
+    """Add auto-start argument to group"""
+    group.add_argument('--auto-start', required=required, dest='auto_start', action='store_true',
                        default=False, help='automatically start')
 
 def __remote(group, required):
@@ -160,28 +160,28 @@ def __remote(group, required):
                        default=False, help='remote copy')
 
 def __combined_output(group):
-    """Add combined_output argument to group"""
-    group.add_argument('--combined_output', dest='combined_output', action='store_true',
+    """Add combined-output argument to group"""
+    group.add_argument('--combined-output', dest='combined_output', action='store_true',
                        help='combine all of the outputs into one output')
 
 def __generate_matrix(group):
-    """Add generate_matrix argument to group"""
-    group.add_argument('--generate_matrix', dest='generate_matrix', action='store_true',
+    """Add generate-matrix argument to group"""
+    group.add_argument('--generate-matrix', dest='generate_matrix', action='store_true',
                        help='generate cgmlst matrix')
 
 def __save_dbs(group):
     """Save all intermediary dbs created for TBProfiler db convergence"""
-    group.add_argument('--save_dbs', dest='save_dbs', action='store_true',
+    group.add_argument('--save-dbs', dest='save_dbs', action='store_true',
                        help='save all intermediary dbs created for TBProfiler db convergence')
 
 def __sample_sheet(group, required):
-    """Add sample_sheet argument to group"""
-    group.add_argument('--sample_sheet', required=required, dest='sample_sheet',
+    """Add sample-sheet argument to group"""
+    group.add_argument('--sample-sheet', required=required, dest='sample_sheet',
                        action='store_true', help='sample sheet input')
 
 def __alter_sample_id(group, required):
-    """Add sample_sheet argument to group"""
-    group.add_argument('--alter_sample_id', required=required,
+    """Add alter-sample-id argument to group"""
+    group.add_argument('--alter-sample-id', required=required,
                        dest='alter_sample_id', action='store_true', default=False,
                        help='alter sample id to be lims ID + sequencing run')
 
@@ -196,6 +196,8 @@ def __help(group):
 def get_main_parser():
     """Get/build the main argument parser"""
     main_parser = argparse.ArgumentParser(prog='jasentool', conflict_handler='resolve')
+    main_parser.add_argument('-v', '--verbose', action='store_true', default=False,
+                             help='enable debug logging')
     sub_parsers = main_parser.add_subparsers(help='--', dest='subparser_name')
     with subparser(sub_parsers, 'find', 'Find sample from given mongo db') as parser:
         with mutex_group(parser, required=True) as group:
@@ -211,19 +213,7 @@ def get_main_parser():
             __prefix(group)
             __help(group)
 
-    with subparser(sub_parsers, 'insert', 'Insert sample(s) into db') as parser:
-        with mutex_group(parser, required=True) as group:
-            __input_file(group, required=False, help='path to json file to be inserted into db')
-            __input_dir(group, required=False, help='path to directory containing sample files')
-        with arg_group(parser, 'required named arguments') as group:
-            __db_name(group, required=True)
-            __db_collection(group, required=True)
-        with arg_group(parser, 'optional arguments') as group:
-            __combined_output(group)
-            __uri(group)
-            __help(group)
-
-    with subparser(sub_parsers, 'validate', 'Compare results from new pipeline to old results') as parser:
+    with subparser(sub_parsers, 'validate-pipelines', 'Compare results from new pipeline to old results') as parser:
         with mutex_group(parser, required=True) as group:
             __input_file(group, required=False, help='input filepath(s)')
             __input_dir(group, required=False, help='path to directory containing sample files')
@@ -240,7 +230,7 @@ def get_main_parser():
             __prefix(group)
             __help(group)
 
-    with subparser(sub_parsers, 'missing', 'Find missing sample data from old runs') as parser:
+    with subparser(sub_parsers, 'identify-missing', 'Find missing sample data from old runs') as parser:
         with arg_group(parser, 'required named arguments') as group:
             __output_file(group, required=True, help='path to mongo db output file')
             __db_name(group, required=True)
@@ -256,7 +246,7 @@ def get_main_parser():
             __alter_sample_id(group, required=False)
             __help(group)
 
-    with subparser(sub_parsers, 'convert', 'Convert file format') as parser:
+    with subparser(sub_parsers, 'transform-file-format', 'Convert file format') as parser:
         with arg_group(parser, 'required named arguments') as group:
             __input_file(group, required=True, help='path to targets tsv file')
             __output_file(group, required=True, help='path to mongo db output file')
@@ -265,7 +255,7 @@ def get_main_parser():
             __accession(group, required=False)
             __help(group)
 
-    with subparser(sub_parsers, 'fix', 'Fix bjorn microbiology csv file') as parser:
+    with subparser(sub_parsers, 'reformat-csv', 'Fix bjorn microbiology csv file') as parser:
         with arg_group(parser, 'required named arguments') as group:
             __csv_file(group, required=True, help='path to bjorn csv file')
             __output_file(group, required=True, help='path to fixed output csv file')
@@ -278,13 +268,13 @@ def get_main_parser():
             __alter_sample_id(group, required=False)
             __help(group)
 
-    with subparser(sub_parsers, 'converge', 'Converge TB mutation catalogues') as parser:
+    with subparser(sub_parsers, 'converge-catalogues', 'Converge TB mutation catalogues') as parser:
         with arg_group(parser, 'optional arguments') as group:
             __output_dir(group, required=False)
             __save_dbs(group)
             __help(group)
 
-    with subparser(sub_parsers, 'qc', 'Run qc on bwa alignment') as parser:
+    with subparser(sub_parsers, 'post-align-qc', 'Run qc on bwa alignment') as parser:
         with arg_group(parser, 'required named arguments') as group:
             __sample_id(group, required=True)
             __bam_file(group, required=True)
@@ -294,6 +284,14 @@ def get_main_parser():
             __bed_file(group, required=False)
             __baits_file(group, required=False)
             __cpus(group)
+            __help(group)
+
+    with subparser(sub_parsers, 'count-reads', 'Count reads in FASTQ file(s)') as parser:
+        with arg_group(parser, 'required named arguments') as group:
+            __input_file(group, required=True, help='path to FASTQ file(s); provide R1 only or R1 + R2')
+            __output_file(group, required=True, help='path to JSON output file')
+        with arg_group(parser, 'optional arguments') as group:
+            __sample_id(group, required=False)
             __help(group)
 
     return main_parser